requirements.txt
flask==2.0.1
werkzeug==2.0.1
grpcio==1.42.0
protobuf==3.19.4
yagrc>=1.1.1
paho-mqtt>=1.5.1
influxdb>=5.3.1
influxdb_client>=1.23.0
pypng>=0.0.20
typing-extensions>=4.3.0
croniter>=1.0.1
python-dateutil>=2.7.0
gunicorn==20.1.0
click>=7.1.2
Jinja2>=3.0
itsdangerous>=2.0
pyngrok==7.0.3
python-dotenv>=1.0.1
firebase-admin==5.3.0
